Add measured concept
Measured defines a relationship between our data and the monoid used to
measure the values in the tree. Tree branches accumulate and cache the measure
of their children via the monoid operation.
Examples are Size and Min, where size's measure counts the elements (1 for any type) and
op is just addition on ints, and where min's measure lifts the value into the monoid, and
the op is less than.